// src/app/enums/EffetEnum.ts
export enum EffetEnum {
  VAMPIRISME = "VAMPIRISME",
  CANNIBALE = "CANNIBALE",
  ESPRIT_EQUIPE = "ESPRIT_EQUIPE",
  CAPITAINE = "CAPITAINE",
  SYMBIOSE = "SYMBIOSE",
  SANG_PUR = "SANG_PUR",
  TSUNAMI = "TSUNAMI",
  HEROISME = "HEROISME",
  IMMUNISE = "IMMUNISE",
  SACRIFICE = "SACRIFICE",
  RESET = "RESET",
  FUSION = "FUSION",
  ELECTROCUTION = "ELECTROCUTION",
  SABOTAGE = "SABOTAGE",
  SABORUPTION = "SABORUPTION",
  INSENSIBLE = "INSENSIBLE",
  SERVIABLE = "SERVIABLE",
  BOUCLIER = "BOUCLIER",
  RECYCLAGE = "RECYCLAGE",
  CORRUPTION = "CORRUPTION",
  POSSESSION = "POSSESSION",
  SILENCE = "SILENCE",
  SAUVETAGE = "SAUVETAGE",
  TROC = "TROC",
  CASSEMURAILLE = "CASSEMURAILLE",
  RENFORT = "RENFORT",
  DOMINATION = "DOMINATION",
  RESURRECTION = "RESURRECTION",
  FIDELITE = "FIDELITE",
  IMPOSTEUR = "IMPOSTEUR",
  PROTECTEURFORET = "PROTECTEURFORET",
  CONVERSION = "CONVERSION",
  RESISTANCE = "RESISTANCE",
  PRISON = "PRISON",
  VISION = "VISION",
  ESPION = "ESPION",
  MEUTE = "MEUTE",
  ILLUMINATI = "ILLUMINATI",
  SECTE = "SECTE",
  CRUAUTE = "CRUAUTE",
  CITADELLE = "CITADELLE",
  NUEE = "NUEE",
  TERREUR = "TERREUR",
  HERITAGE = "HERITAGE",
  SOUTIEN = "SOUTIEN",
  CRYPTE = "CRYPTE",
  AMITIE = "AMITIE",
  ENTERREMENT = "ENTERREMENT",
  DUOTERREMENT = "DUOTERREMENT",
  PALISSADE = "PALISSADE",
  SURVIVANT = "SURVIVANT",
  SABOTEUR = "SABOTEUR",
  POISSON = "POISSON",
  TRAHISON = "TRAHISON",
  STOP = "STOP",
  MENTALISME= "MENTALISME",
  TARDIF = "TARDIF",
  DEVOREUR = "DEVOREUR",
  ABSORPTION = "ABSORPTION",
  VOIX = "VOIX",
  MEURTRE = "MEURTRE",
  CAUCHEMARD = "CAUCHEMARD",
  CHROPIE = "CHROPIE",
  MATINAL = "MATINAL",
  PARI = "PARI",
  RESISTANCE_INSTANT = "RESISTANCE_INSTANT",
  SEPT = "SEPT",
  SIX = "SIX",
  BOUCLIERPLUS = "BOUCLIERPLUS",
  SABOTEURPLUS = "SABOTEURPLUS",
  CINQ = "CINQ",
  QUATRE = "QUATRE",
  INSENSIBLEPLUS = "INSENSIBLEPLUS",
  MELEE = 'MELEE',
  SECOND = 'SECOND',
  TROISIEME = 'TROISIEME',
  EGOISME = 'EGOISME',
  KAMIKAZE = 'KAMIKAZE'

}